Make TestableTarget conform ModelConvertible
